refactor(logging): move interface for logging into logging subpackage
