--- src/common/index.ts
@@ -1,1 +1,0 @@
-export type { Logger, LogArgs, LoggerBindings } from "./types";
