+++ src/logging/index.tsx
@@ -1,2 +1,2 @@
-export { type Logger, LOG_LEVEL } from "./types";
+export { type Logger, type LogArgs, type LoggerBindings, LOG_LEVEL } from "./types";
 export { createFileLogger } from "./file-logger";
